Fix Image.Size type: API returns float, not int64
OVH API /cloud/project/{sn}/image returns image sizes as floats
(e.g. 9.9990234375 GB). The int64 type was causing JSON unmarshal
errors on real API responses.

+++ pkg/ovh/types.go
@@ -92,8 +92,8 @@ type Image struct {
 	Region       string `json:"region"`
 	CreationDate string `json:"creationDate"`
 	MinRAM       int    `json:"minRam"`
-	MinDisk      int    `json:"minDisk"`
-	Size         int64  `json:"size"`
+	MinDisk      int     `json:"minDisk"`
+	Size         float64 `json:"size"`
 	Type         string `json:"type"`
 	User         string `json:"user"`
 	Visibility   string `json:"visibility"`
